Split and move files into Bindings module
The binding.rb file was getting to long so the classes have been
split into seperate files and stored in a Binding module. The binding
is now built using the module method, Binding.build.
