+++ src/binding.rb
@@ -1,0 +1,37 @@
+# frozen_string_literal: true
+
+#==============================================================================
+# Copyright (C) 2017 Stephen F. Norledge and Alces Software Ltd.
+#
+# This file/package is part of Alces Metalware.
+#
+# Alces Metalware is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License
+# as published by the Free Software Foundation, either version 3 of
+# the License, or (at your option) any later version.
+#
+# Alces Metalware is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this package.  If not, see <http://www.gnu.org/licenses/>.
+#
+# For more information on the Alces Metalware, please visit:
+# https://github.com/alces-software/metalware
+#==============================================================================
+
+# Dynamically requires all ruby files in binding directory
+Dir[File.join(File.dirname(__FILE__), 'binding', '**/*.rb')].map do |file|
+  require file.sub("#{File.dirname(__FILE__)}/", '').chomp('.rb')
+end
+
+module Metalware
+  module Binding
+    def self.build(config, node = nil)
+      alces_binding = Parameter.new(config: config, node_name: node)
+      Wrapper.new(alces_binding).get_binding
+    end
+  end
+end
